Preserve cell indices when exporting a network to NeuroML.
When exporting to NeuroML, the "id" attribute of the <instance/>
element describing a cell position in the exported file is supposed to
correspond to neuroConstruct's "cell number" value for that cell - a
unique identifier of the cell within its group.

The problem here was that the "id" attribute was instead set to the
index indicating the order in which the cell's position record had
been added to the array of generated cell positions; this essentially
results in a permutation of the cell identities within each group if
this index doesn't correspond to the "cell number" value. Usually this
is not a problem, as "cell number"s are generated sequentially as cells
are generated, making the two indices coincide; but it becomes
apparent if manually generating cells in a nontrivial order through
the python scripting interface.